add AutoLEDShutoff project
This project is a simple test to show how to turn the LEDs off when the battery is low, and waits until power is disconnected before powering them again.
This is a basic state machine, but a more advanced one needs to be created for
the 'real' lights. This project was created for rev 4, so it will need to be updated for later revisions.